+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - OpenAI Compatible: Add support for using Responses API via `responses_api` model arg.
 - Eval Set: Add `eval_set_id` to log file (unique id for eval set across invocations for the same `log_dir`).
 - Task Display: Render tool outputs based on model events rather than tool events (compatibile w/ external tool calling e.g. for agent bridge).
+- Sandbox Service: Don't emit sandbox events for sandbox service RPC calls.
 - Hooks: New `EvalSetStart` and `EvalSetEnd` hook methods.
 - Scoring: `inspect score` now supports streaming via the `--stream` argument.
 - Inspect View: Starting the view server with a path to a specific log file will automatically open that log file (if it exists) rather than showing the log list.